New version of Semplicemente - 1.9.3

+++ footer.php
@@ -9,21 +9,22 @@
 ?>
 
 	</div><!-- #content -->
-
-	<footer id="colophon" class="site-footer">
-		<div class="site-info">
-			<?php
-			$copyrightText = get_theme_mod('semplicemente_theme_options_copyrighttext', '&copy; '.date('Y').' '. get_bloginfo('name'));
-			if ($copyrightText || is_customize_preview()): ?>
-				<span class="custom"><?php echo wp_kses($copyrightText, semplicemente_allowed_html()); ?></span>
-			<?php endif; ?>
-			<span class="sep"> | </span>
-			<?php
-			/* translators: 1: theme name, 2: theme developer */
-			printf( esc_html__( 'WordPress Theme: %1$s by %2$s.', 'semplicemente' ), '<a title="Semplicemente Theme" href="https://crestaproject.com/downloads/semplicemente/" rel="nofollow" target="_blank">Semplicemente</a>', 'CrestaProject' );
-			?>
-		</div><!-- .site-info -->
-	</footer><!-- #colophon -->
+	<?php if ( ! function_exists( 'elementor_theme_do_location' ) || ! elementor_theme_do_location( 'footer' ) ) : ?>
+		<footer id="colophon" class="site-footer">
+			<div class="site-info">
+				<?php
+				$copyrightText = get_theme_mod('semplicemente_theme_options_copyrighttext', '&copy; '.date('Y').' '. get_bloginfo('name'));
+				if ($copyrightText || is_customize_preview()): ?>
+					<span class="custom"><?php echo wp_kses($copyrightText, semplicemente_allowed_html()); ?></span>
+				<?php endif; ?>
+				<span class="sep"> | </span>
+				<?php
+				/* translators: 1: theme name, 2: theme developer */
+				printf( esc_html__( 'WordPress Theme: %1$s by %2$s.', 'semplicemente' ), '<a title="Semplicemente Theme" href="https://crestaproject.com/downloads/semplicemente/" rel="nofollow" target="_blank">Semplicemente</a>', 'CrestaProject' );
+				?>
+			</div><!-- .site-info -->
+		</footer><!-- #colophon -->
+	<?php endif; ?>
 </div><!-- #page -->
 
 <?php wp_footer(); ?>
